move cap string parsing to core & add to ad4m-connect
