+++ core/src/index.ts
@@ -25,3 +25,4 @@ export * from "./neighbourhood/Neighbourhood";
 export * from "./neighbourhood/NeighbourhoodProxy";
 export * from "./typeDefs";
 export * from "./DID";
+export * from "./utils";
